requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.3
pandas==2.0.3
Pillow==9.4.0
Pillow==10.0.1
Requests==2.31.0
selenium==4.12.0
streamlit==1.27.0
yfinance==0.2.28
